feat: add deeplink parsing for login SDK intents
